requirements.txt
streamlit 
langchain-google-genai 
python-dotenv
fpdf
pandas
uuid